pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Praneeth Yerrapragada, Patrick Alexis" ]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
aiostream = "^0.5.2"
llama-index = "0.10.28"
llama-index-core = "0.10.28"
llama-index-vector-stores-pinecone = "^0.1.7"
traceloop-sdk = "^0.19.0"
langchain = "^0.2.1"
langchain-openai = "^0.1.8"
openai = "^1.30.5"
tenacity = "^8.3.0"
rapidfuzz = "^3.9.2"
pydantic = "^2.7.2"
dateparser = "^1.2.0"
pandas = "^2.2.2"
path = "^16.14.0"
pytest-xdist = "^3.6.1"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
asyncio = "^3.4.3"
pytest-postgresql = "^6.0.0"
psycopg-binary = "^3.1.19"
python-multipart = "^0.0.9"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.sqlalchemy]
extras = [ "asyncio" ]
version = "^2.0.30"

[tool.poetry.dependencies.llama-index-agent-openai]
version = "0.2.2"

[tool.black]
line-length = 119

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"